add react overview. add docs templates.
