+++ community/envs/community-react/community-react.main.runtime.ts
@@ -5,7 +5,7 @@ import { EnvsAspect, EnvsMain } from '@teambit/envs';
 import { CommunityReactAspect } from './community-react.aspect';
 import { myReactTemplate } from './templates/my-react-template';
 import { myEntityTemplate } from './templates/my-entity-template';
-import { mdxCommunity } from './templates/mdx-component';
+import { bitDevDocs } from './templates/bitdev-docs';
 import { transformTsConfig } from './typescript/transform-tsconfig';
 
 // import { previewConfigTransformer, devServerConfigTransformer } from './webpack/webpack-transformers';
@@ -92,7 +92,7 @@ export class CommunityReactMain {
     ]);
     envs.registerEnv(templatesReactEnv);
 
-    generator.registerComponentTemplate([myReactTemplate, myEntityTemplate, mdxCommunity]);
+    generator.registerComponentTemplate([myReactTemplate, myEntityTemplate, bitDevDocs]);
 
     return new CommunityReactMain();
   }
