+++ community/envs/community-react/templates/bitdev-docs/files/composition-file.ts
@@ -1,20 +1,30 @@
-import React from 'react';
+import { ComponentContext } from '@teambit/generator';
+
+export function compositionFile({ namePascalCase }: ComponentContext) {
+  return `import React from 'react';
 import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
 import { MDXLayout } from '@teambit/mdx.ui.mdx-layout';
 import { BasicHeader } from '@teambit/community.ui.header.header/dist/header.composition';
 import { WideColumn } from '@teambit/base-ui.layout.page-frame';
 import { Docs } from '@teambit/docs.ui.docs';
-import { ReactOverview } from './index';
+import { ${namePascalCase} } from './index';
 
 const routes = [
   {
     path: '',
     title: '',
-    component: <ReactOverview />,
+    component: <${namePascalCase} />,
   },
 ];
 
 const DocsApp = () => {
+  const routes = [
+    {
+      path: '',
+      title: '',
+      component: <${namePascalCase} />,
+    },
+  ];
   return <Docs baseUrl="" routes={routes} />;
 };
 
@@ -29,12 +39,13 @@ export const DisplayedInBitDev = () => {
   );
 };
 
-export const BasicReactOverview = () => {
+export const Basic${namePascalCase} = () => {
   return (
     <ThemeCompositions>
       <MDXLayout>
-        <ReactOverview />
+        <${namePascalCase} />
       </MDXLayout>
     </ThemeCompositions>
   );
-};
+};`;
+}
