+++ community/envs/community-react/templates/bitdev-docs/files/implementation-file.ts
