+++ community/envs/community-react/templates/bitdev-docs/files/index-file.ts
@@ -1,5 +1,5 @@
 import { ComponentContext } from '@teambit/generator';
 
 export function indexFile({ name, namePascalCase }: ComponentContext) {
-  return `export { default as ${namePascalCase}, default } from './${name}.mdx`;
+  return `export { default as ${namePascalCase}, default } from './${name}.mdx'`;
 }
