--- community/envs/community-react/templates/mdx-component/files/composition-file.ts
@@ -1,18 +1,0 @@
-import { ComponentContext } from '@teambit/generator';
-
-export function compositionFile({ namePascalCase }: ComponentContext) {
-  return `import React from 'react';
-import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
-import { MDXLayout } from '@teambit/mdx.ui.mdx-layout';
-import { ${namePascalCase} } from './index';
-
-export const Basic${namePascalCase} = () => {
-  return (
-    <ThemeCompositions>
-      <MDXLayout>
-        <${namePascalCase} />
-      </MDXLayout>
-    </ThemeCompositions>
-  );
-};`;
-}
