+++ react/content/extending-react/extending-react.composition.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
 import { MDXLayout } from '@teambit/mdx.ui.mdx-layout';
+import { Header } from '@teambit/community.ui.header.header';
+import { CommunityDocs } from '@teambit/docs.ui.community-docs';
 import { ExtendingReact } from './index';
 
 export const BasicExtendingReact = () => {
@@ -9,6 +11,17 @@ export const BasicExtendingReact = () => {
       <MDXLayout>
         <ExtendingReact />
       </MDXLayout>
+    </ThemeCompositions>
+  );
+};
+
+export const ExtendingReactInDocsApp = () => {
+  return (
+    <ThemeCompositions>
+      <Header />
+      <CommunityDocs routes={[]}>
+        <ExtendingReact />
+      </CommunityDocs>
     </ThemeCompositions>
   );
 };
