+++ react/content/extending-react/temp.composition.tsx
@@ -1,40 +1,28 @@
 import React from 'react';
-import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
-import { MDXLayout } from '@teambit/mdx.ui.mdx-layout';
 import { BasicHeader } from '@teambit/community.ui.header.header/dist/header.composition';
 import { WideColumn } from '@teambit/base-ui.layout.page-frame';
+import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
 import { Docs } from '@teambit/docs.ui.docs';
-import { ReactOverview } from './index';
-
-const routes = [
-  {
-    path: '',
-    title: '',
-    component: <ReactOverview />,
-  },
-];
+import { ExtendingReact } from './index';
 
 const DocsApp = () => {
+  const routes = [
+    {
+      path: '',
+      title: '',
+      component: <ExtendingReact />,
+    },
+  ];
   return <Docs baseUrl="" routes={routes} />;
 };
 
 export const DisplayedInBitDev = () => {
   return (
-    <>
+    <ThemeCompositions>
       <BasicHeader />
       <WideColumn>
         <DocsApp />
       </WideColumn>
-    </>
-  );
-};
-
-export const BasicReactOverview = () => {
-  return (
-    <ThemeCompositions>
-      <MDXLayout>
-        <ReactOverview />
-      </MDXLayout>
     </ThemeCompositions>
   );
 };
