+++ react/content/react-overview/index.ts
@@ -1,1 +1,1 @@
-export { default as ReactOverview, default } from './react-overview.mdx';
+export { default as ReactOverview, default } from './react-overview.mdx'
