+++ react/content/react-overview/react-overview.docs.mdx
@@ -1,9 +1,9 @@
 ---
-description: 'An MDX component.'
-labels: ['content', 'mdx']
+description: 'A community doc.'
+labels: ['content', 'mdx',]
 ---
 
-import { ReactOverview } from './index';
+import ReactOverview from './index';
 
 ### Component usage
 ```js
@@ -14,7 +14,7 @@ import { ReactOverview } from './index';
 
 ```js
 import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions'
-import { MDXLayout } from '@teambit/ui.mdx-layout'
+import { MDXLayout } from '@teambit/mdx.ui.mdx-layout'
 import { ReactOverview } from './index'
 
 <ThemeCompositions>
