+++ react/content/react-router/react-router.docs.mdx
@@ -6,6 +6,7 @@ labels: ['content', 'mdx']
 import { ReactRouter } from './index';
 
 ### Component usage
+
 ```js
 <ReactRouter />
 ```
@@ -13,13 +14,13 @@ import { ReactRouter } from './index';
 ### Render with theme and MDX providers
 
 ```js
-import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions'
-import { MDXLayout } from '@teambit/mdx.ui.mdx-layout'
-import { ReactRouter } from './index'
+import { ThemeCompositions } from '@teambit/documenter.theme.theme-compositions';
+import { MDXLayout } from '@teambit/mdx.ui.mdx-layout';
+import { ReactRouter } from './index';
 
 <ThemeCompositions>
   <MDXLayout>
     <ReactRouter />
   </MDXLayout>
-</ThemeCompositions>
+</ThemeCompositions>;
 ```
